refactor parser ts
