--- src/parser/_parser.ts
@@ -1,671 +1,0 @@
-//@ts-nocheck
-import { Color } from '../color';
-import { commaWsp, DEFAULT_SVG_FONT_SIZE, reNum, svgNS } from '../constants';
-import { FabricObject } from '../shapes/object.class';
-import { cos, degreesToRadians, multiplyTransformMatrices, parseUnit, sin, toFixed } from '../util';
-import { reFontDeclaration, attributesMap, colorAttributes, svgValidParentsRegEx, cPath, fSize, cssRules, svgViewBoxElementsRegEx, reViewBoxAttrValue } from './constants';
-import { parseStyleAttribute } from './parseStyleAttribute';
-
-export function parseFontDeclaration(value, oStyle) {
-    var match = value.match(reFontDeclaration);
-
-    if (!match) {
-        return;
-    }
-    var fontStyle = match[1],
-        // font variant is not used
-        // fontVariant = match[2],
-        fontWeight = match[3], fontSize = match[4], lineHeight = match[5], fontFamily = match[6];
-
-    if (fontStyle) {
-        oStyle.fontStyle = fontStyle;
-    }
-    if (fontWeight) {
-        oStyle.fontWeight = isNaN(parseFloat(fontWeight)) ? fontWeight : parseFloat(fontWeight);
-    }
-    if (fontSize) {
-        oStyle.fontSize = parseUnit(fontSize);
-    }
-    if (fontFamily) {
-        oStyle.fontFamily = fontFamily;
-    }
-    if (lineHeight) {
-        oStyle.lineHeight = lineHeight === 'normal' ? 1 : lineHeight;
-    }
-}
-export function normalizeAttr(attr) {
-    // transform attribute names
-    if (attr in attributesMap) {
-        return attributesMap[attr];
-    }
-    return attr;
-}
-export function normalizeValue(attr, value, parentAttributes, fontSize) {
-    var isArray = Array.isArray(value), parsed;
-
-    if ((attr === 'fill' || attr === 'stroke') && value === 'none') {
-        value = '';
-    }
-    else if (attr === 'strokeUniform') {
-        return (value === 'non-scaling-stroke');
-    }
-    else if (attr === 'strokeDashArray') {
-        if (value === 'none') {
-            value = null;
-        }
-        else {
-            value = value.replace(/,/g, ' ').split(/\s+/).map(parseFloat);
-        }
-    }
-    else if (attr === 'transformMatrix') {
-        if (parentAttributes && parentAttributes.transformMatrix) {
-            value = multiplyTransformMatrices(
-                parentAttributes.transformMatrix, parseTransformAttribute(value));
-        }
-        else {
-            value = parseTransformAttribute(value);
-        }
-    }
-    else if (attr === 'visible') {
-        value = value !== 'none' && value !== 'hidden';
-        // display=none on parent element always takes precedence over child element
-        if (parentAttributes && parentAttributes.visible === false) {
-            value = false;
-        }
-    }
-    else if (attr === 'opacity') {
-        value = parseFloat(value);
-        if (parentAttributes && typeof parentAttributes.opacity !== 'undefined') {
-            value *= parentAttributes.opacity;
-        }
-    }
-    else if (attr === 'textAnchor' /* text-anchor */) {
-        value = value === 'start' ? 'left' : value === 'end' ? 'right' : 'center';
-    }
-    else if (attr === 'charSpacing') {
-        // parseUnit returns px and we convert it to em
-        parsed = parseUnit(value, fontSize) / fontSize * 1000;
-    }
-    else if (attr === 'paintFirst') {
-        var fillIndex = value.indexOf('fill');
-        var strokeIndex = value.indexOf('stroke');
-        var value = 'fill';
-        if (fillIndex > -1 && strokeIndex > -1 && strokeIndex < fillIndex) {
-            value = 'stroke';
-        }
-        else if (fillIndex === -1 && strokeIndex > -1) {
-            value = 'stroke';
-        }
-    }
-    else if (attr === 'href' || attr === 'xlink:href' || attr === 'font') {
-        return value;
-    }
-    else if (attr === 'imageSmoothing') {
-        return (value === 'optimizeQuality');
-    }
-    else {
-        parsed = isArray ? value.map(parseUnit) : parseUnit(value, fontSize);
-    }
-
-    return (!isArray && isNaN(parsed) ? value : parsed);
-}
-/**
-  * @private
-  */
-export function getSvgRegex(arr) {
-    return new RegExp('^(' + arr.join('|') + ')\\b', 'i');
-}
-/**
- * @private
- * @param {Object} attributes Array of attributes to parse
- */
-export function setStrokeFillOpacity(attributes) {
-    for (var attr in colorAttributes) {
-
-        if (typeof attributes[colorAttributes[attr]] === 'undefined' || attributes[attr] === '') {
-            continue;
-        }
-
-        if (typeof attributes[attr] === 'undefined') {
-            if (!FabricObject.prototype[attr]) {
-                continue;
-            }
-            attributes[attr] = FabricObject.prototype[attr];
-        }
-
-        if (attributes[attr].indexOf('url(') === 0) {
-            continue;
-        }
-
-        var color = new Color(attributes[attr]);
-        attributes[attr] = color.setAlpha(toFixed(color.getAlpha() * attributes[colorAttributes[attr]], 2)).toRgba();
-    }
-    return attributes;
-}
-/**
- * @private
- */
-export function _getMultipleNodes(doc, nodeNames) {
-    var nodeName, nodeArray = [], nodeList, i, len;
-    for (i = 0, len = nodeNames.length; i < len; i++) {
-        nodeName = nodeNames[i];
-        nodeList = doc.getElementsByTagName(nodeName);
-        nodeArray = nodeArray.concat(Array.prototype.slice.call(nodeList));
-    }
-    return nodeArray;
-}
-
-export function parsePointsAttribute(points) {
-
-    // points attribute is required and must not be empty
-    if (!points) {
-        return null;
-    }
-
-    // replace commas with whitespace and remove bookending whitespace
-    points = points.replace(/,/g, ' ').trim();
-
-    points = points.split(/\s+/);
-    var parsedPoints = [], i, len;
-
-    for (i = 0, len = points.length; i < len; i += 2) {
-        parsedPoints.push({
-            x: parseFloat(points[i]),
-            y: parseFloat(points[i + 1])
-        });
-    }
-
-    // odd number of points is an error
-    // if (parsedPoints.length % 2 !== 0) {
-    //   return null;
-    // }
-    return parsedPoints;
-}
-export function rotateMatrix(matrix, args) {
-    var cosValue = cos(args[0]), sinValue = sin(args[0]), x = 0, y = 0;
-    if (args.length === 3) {
-        x = args[1];
-        y = args[2];
-    }
-
-    matrix[0] = cosValue;
-    matrix[1] = sinValue;
-    matrix[2] = -sinValue;
-    matrix[3] = cosValue;
-    matrix[4] = x - (cosValue * x - sinValue * y);
-    matrix[5] = y - (sinValue * x + cosValue * y);
-}
-export function scaleMatrix(matrix, args) {
-    var multiplierX = args[0], multiplierY = (args.length === 2) ? args[1] : args[0];
-
-    matrix[0] = multiplierX;
-    matrix[3] = multiplierY;
-}
-export function skewMatrix(matrix, args, pos) {
-    matrix[pos] = Math.tan(degreesToRadians(args[0]));
-}
-export function translateMatrix(matrix, args) {
-    matrix[4] = args[0];
-    if (args.length === 2) {
-        matrix[5] = args[1];
-    }
-}
-// == begin transform regexp
-let number = reNum, skewX = '(?:(skewX)\\s*\\(\\s*(' + number + ')\\s*\\))', skewY = '(?:(skewY)\\s*\\(\\s*(' + number + ')\\s*\\))', rotate = '(?:(rotate)\\s*\\(\\s*(' + number + ')(?:' +
-    commaWsp + '(' + number + ')' +
-    commaWsp + '(' + number + '))?\\s*\\))', scale = '(?:(scale)\\s*\\(\\s*(' + number + ')(?:' +
-        commaWsp + '(' + number + '))?\\s*\\))', translate = '(?:(translate)\\s*\\(\\s*(' + number + ')(?:' +
-            commaWsp + '(' + number + '))?\\s*\\))', matrix = '(?:(matrix)\\s*\\(\\s*' +
-                '(' + number + ')' + commaWsp +
-                '(' + number + ')' + commaWsp +
-                '(' + number + ')' + commaWsp +
-                '(' + number + ')' + commaWsp +
-                '(' + number + ')' + commaWsp +
-                '(' + number + ')' +
-                '\\s*\\))', transform = '(?:' +
-                    matrix + '|' +
-                    translate + '|' +
-                    scale + '|' +
-                    rotate + '|' +
-                    skewX + '|' +
-                    skewY +
-                    ')', transforms = '(?:' + transform + '(?:' + commaWsp + '*' + transform + ')*' + ')', transformList = '^\\s*(?:' + transforms + '?)\\s*$',
-    // http://www.w3.org/TR/SVG/coords.html#TransformAttribute
-    reTransformList = new RegExp(transformList),
-    // == end transform regexp
-    reTransform = new RegExp(transform, 'g');
-export function parseTransformAttribute(attributeValue) {
-
-    // start with identity matrix
-    var matrix = iMatrix.concat(), matrices = [];
-
-    // return if no argument was given or
-    // an argument does not match transform attribute regexp
-    if (!attributeValue || (attributeValue && !reTransformList.test(attributeValue))) {
-        return matrix;
-    }
-
-    attributeValue.replace(reTransform, function (match) {
-
-        var m = new RegExp(transform).exec(match).filter(function (match) {
-            // match !== '' && match != null
-            return (!!match);
-        }), operation = m[1], args = m.slice(2).map(parseFloat);
-
-        switch (operation) {
-            case 'translate':
-                translateMatrix(matrix, args);
-                break;
-            case 'rotate':
-                args[0] = degreesToRadians(args[0]);
-                rotateMatrix(matrix, args);
-                break;
-            case 'scale':
-                scaleMatrix(matrix, args);
-                break;
-            case 'skewX':
-                skewMatrix(matrix, args, 2);
-                break;
-            case 'skewY':
-                skewMatrix(matrix, args, 1);
-                break;
-            case 'matrix':
-                matrix = args;
-                break;
-        }
-
-        // snapshot current matrix into matrices array
-        matrices.push(matrix.concat());
-        // reset
-        matrix = iMatrix.concat();
-    });
-
-    var combinedMatrix = matrices[0];
-    while (matrices.length > 1) {
-        matrices.shift();
-        combinedMatrix = multiplyTransformMatrices(combinedMatrix, matrices[0]);
-    }
-    return combinedMatrix;
-}
-/**
- * Returns an object of attributes' name/value, given element and an array of attribute names;
- * Parses parent "g" nodes recursively upwards.
- * @param {DOMElement} element Element to parse
- * @param {Array} attributes Array of attributes to parse
- * @return {Object} object containing parsed attributes' names/values
- */
-export function parseAttributes(element, attributes, svgUid) {
-
-    if (!element) {
-        return;
-    }
-
-    var value, parentAttributes = {}, fontSize, parentFontSize;
-
-    if (typeof svgUid === 'undefined') {
-        svgUid = element.getAttribute('svgUid');
-    }
-    // if there's a parent container (`g` or `a` or `symbol` node), parse its attributes recursively upwards
-    if (element.parentNode && svgValidParentsRegEx.test(element.parentNode.nodeName)) {
-        parentAttributes = parseAttributes(element.parentNode, attributes, svgUid);
-    }
-
-    var ownAttributes = attributes.reduce(function (memo, attr) {
-        value = element.getAttribute(attr);
-        if (value) { // eslint-disable-line
-            memo[attr] = value;
-        }
-        return memo;
-    }, {});
-    // add values parsed from style, which take precedence over attributes
-    // (see: http://www.w3.org/TR/SVG/styling.html#UsingPresentationAttributes)
-    var cssAttrs = Object.assign(
-        getGlobalStylesForElement(element, svgUid),
-        parseStyleAttribute(element)
-    );
-    ownAttributes = Object.assign(
-        ownAttributes,
-        cssAttrs
-    );
-    if (cssAttrs[cPath]) {
-        element.setAttribute(cPath, cssAttrs[cPath]);
-    }
-    fontSize = parentFontSize = parentAttributes.fontSize || DEFAULT_SVG_FONT_SIZE;
-    if (ownAttributes[fSize]) {
-        // looks like the minimum should be 9px when dealing with ems. this is what looks like in browsers.
-        ownAttributes[fSize] = fontSize = parseUnit(ownAttributes[fSize], parentFontSize);
-    }
-
-    var normalizedAttr, normalizedValue, normalizedStyle = {};
-    for (var attr in ownAttributes) {
-        normalizedAttr = normalizeAttr(attr);
-        normalizedValue = normalizeValue(normalizedAttr, ownAttributes[attr], parentAttributes, fontSize);
-        normalizedStyle[normalizedAttr] = normalizedValue;
-    }
-    if (normalizedStyle && normalizedStyle.font) {
-        parseFontDeclaration(normalizedStyle.font, normalizedStyle);
-    }
-    var mergedAttrs = Object.assign(parentAttributes, normalizedStyle);
-    return svgValidParentsRegEx.test(element.nodeName) ? mergedAttrs : setStrokeFillOpacity(mergedAttrs);
-}
-
-/**
- * @private
- */
-export function parseStyleString(style, oStyle) {
-    var attr, value;
-    style.replace(/;\s*$/, '').split(';').forEach(function (chunk) {
-        var pair = chunk.split(':');
-
-        attr = pair[0].trim().toLowerCase();
-        value = pair[1].trim();
-
-        oStyle[attr] = value;
-    });
-}
-
-/**
- * @private
- */
-export function parseStyleObject(style, oStyle) {
-    var attr, value;
-    for (var prop in style) {
-        if (typeof style[prop] === 'undefined') {
-            continue;
-        }
-
-        attr = prop.toLowerCase();
-        value = style[prop];
-
-        oStyle[attr] = value;
-    }
-}
-
-/**
- * @private
- */
-export function getGlobalStylesForElement(element, svgUid) {
-    var styles = {};
-    for (var rule in cssRules[svgUid]) {
-        if (elementMatchesRule(element, rule.split(' '))) {
-            for (var property in cssRules[svgUid][rule]) {
-                styles[property] = cssRules[svgUid][rule][property];
-            }
-        }
-    }
-    return styles;
-}
-
-/**
- * @private
- */
-export function elementMatchesRule(element, selectors) {
-    var firstMatching, parentMatching = true;
-    //start from rightmost selector.
-    firstMatching = selectorMatches(element, selectors.pop());
-    if (firstMatching && selectors.length) {
-        parentMatching = doesSomeParentMatch(element, selectors);
-    }
-    return firstMatching && parentMatching && (selectors.length === 0);
-}
-
-export function doesSomeParentMatch(element, selectors) {
-    var selector, parentMatching = true;
-    while (element.parentNode && element.parentNode.nodeType === 1 && selectors.length) {
-        if (parentMatching) {
-            selector = selectors.pop();
-        }
-        element = element.parentNode;
-        parentMatching = selectorMatches(element, selector);
-    }
-    return selectors.length === 0;
-}
-
-/**
- * @private
- */
-export function selectorMatches(element, selector) {
-    var nodeName = element.nodeName,
-        classNames = element.getAttribute('class'),
-        id = element.getAttribute('id'), matcher, i;
-    // i check if a selector matches slicing away part from it.
-    // if i get empty string i should match
-    matcher = new RegExp('^' + nodeName, 'i');
-    selector = selector.replace(matcher, '');
-    if (id && selector.length) {
-        matcher = new RegExp('#' + id + '(?![a-zA-Z\\-]+)', 'i');
-        selector = selector.replace(matcher, '');
-    }
-    if (classNames && selector.length) {
-        classNames = classNames.split(' ');
-        for (i = classNames.length; i--;) {
-            matcher = new RegExp('\\.' + classNames[i] + '(?![a-zA-Z\\-]+)', 'i');
-            selector = selector.replace(matcher, '');
-        }
-    }
-    return selector.length === 0;
-}
-
-/**
- * @private
- * to support IE8 missing getElementById on SVGdocument and on node xmlDOM
- */
-export function elementById(doc, id) {
-    var el;
-    doc.getElementById && (el = doc.getElementById(id));
-    if (el) {
-        return el;
-    }
-    var node, i, len, nodelist = doc.getElementsByTagName('*');
-    for (i = 0, len = nodelist.length; i < len; i++) {
-        node = nodelist[i];
-        if (id === node.getAttribute('id')) {
-            return node;
-        }
-    }
-}
-
-/**
- * @private
- */
-export function parseUseDirectives(doc) {
-    var nodelist = _getMultipleNodes(doc, ['use', 'svg:use']), i = 0;
-    while (nodelist.length && i < nodelist.length) {
-        var el = nodelist[i],
-            xlinkAttribute = el.getAttribute('xlink:href') || el.getAttribute('href');
-
-        if (xlinkAttribute === null) {
-            return;
-        }
-
-        var xlink = xlinkAttribute.slice(1),
-            x = el.getAttribute('x') || 0,
-            y = el.getAttribute('y') || 0,
-            el2 = elementById(doc, xlink).cloneNode(true),
-            currentTrans = (el2.getAttribute('transform') || '') + ' translate(' + x + ', ' + y + ')',
-            parentNode,
-            oldLength = nodelist.length, attr,
-            j,
-            attrs,
-            len,
-            namespace = svgNS;
-
-        applyViewboxTransform(el2);
-        if (/^svg$/i.test(el2.nodeName)) {
-            var el3 = el2.ownerDocument.createElementNS(namespace, 'g');
-            for (j = 0, attrs = el2.attributes, len = attrs.length; j < len; j++) {
-                attr = attrs.item(j);
-                el3.setAttributeNS(namespace, attr.nodeName, attr.nodeValue);
-            }
-            // el2.firstChild != null
-            while (el2.firstChild) {
-                el3.appendChild(el2.firstChild);
-            }
-            el2 = el3;
-        }
-
-        for (j = 0, attrs = el.attributes, len = attrs.length; j < len; j++) {
-            attr = attrs.item(j);
-            if (attr.nodeName === 'x' || attr.nodeName === 'y' ||
-                attr.nodeName === 'xlink:href' || attr.nodeName === 'href') {
-                continue;
-            }
-
-            if (attr.nodeName === 'transform') {
-                currentTrans = attr.nodeValue + ' ' + currentTrans;
-            }
-            else {
-                el2.setAttribute(attr.nodeName, attr.nodeValue);
-            }
-        }
-
-        el2.setAttribute('transform', currentTrans);
-        el2.setAttribute('instantiated_by_use', '1');
-        el2.removeAttribute('id');
-        parentNode = el.parentNode;
-        parentNode.replaceChild(el2, el);
-        // some browsers do not shorten nodelist after replaceChild (IE8)
-        if (nodelist.length === oldLength) {
-            i++;
-        }
-    }
-}
-
-
-
-/**
- * Add a <g> element that envelop all child elements and makes the viewbox transformMatrix descend on all elements
- */
-export function applyViewboxTransform(element) {
-    if (!svgViewBoxElementsRegEx.test(element.nodeName)) {
-        return {};
-    }
-    var viewBoxAttr = element.getAttribute('viewBox'),
-        scaleX = 1,
-        scaleY = 1,
-        minX = 0,
-        minY = 0,
-        viewBoxWidth, viewBoxHeight, matrix, el,
-        widthAttr = element.getAttribute('width'),
-        heightAttr = element.getAttribute('height'),
-        x = element.getAttribute('x') || 0,
-        y = element.getAttribute('y') || 0,
-        preserveAspectRatio = element.getAttribute('preserveAspectRatio') || '',
-        missingViewBox = (!viewBoxAttr || !(viewBoxAttr = viewBoxAttr.match(reViewBoxAttrValue))),
-        missingDimAttr = (!widthAttr || !heightAttr || widthAttr === '100%' || heightAttr === '100%'),
-        toBeParsed = missingViewBox && missingDimAttr,
-        parsedDim = {}, translateMatrix = '', widthDiff = 0, heightDiff = 0;
-
-    parsedDim.width = 0;
-    parsedDim.height = 0;
-    parsedDim.toBeParsed = toBeParsed;
-
-    if (missingViewBox) {
-        if (((x || y) && element.parentNode && element.parentNode.nodeName !== '#document')) {
-            translateMatrix = ' translate(' + parseUnit(x) + ' ' + parseUnit(y) + ') ';
-            matrix = (element.getAttribute('transform') || '') + translateMatrix;
-            element.setAttribute('transform', matrix);
-            element.removeAttribute('x');
-            element.removeAttribute('y');
-        }
-    }
-
-    if (toBeParsed) {
-        return parsedDim;
-    }
-
-    if (missingViewBox) {
-        parsedDim.width = parseUnit(widthAttr);
-        parsedDim.height = parseUnit(heightAttr);
-        // set a transform for elements that have x y and are inner(only) SVGs
-        return parsedDim;
-    }
-    minX = -parseFloat(viewBoxAttr[1]);
-    minY = -parseFloat(viewBoxAttr[2]);
-    viewBoxWidth = parseFloat(viewBoxAttr[3]);
-    viewBoxHeight = parseFloat(viewBoxAttr[4]);
-    parsedDim.minX = minX;
-    parsedDim.minY = minY;
-    parsedDim.viewBoxWidth = viewBoxWidth;
-    parsedDim.viewBoxHeight = viewBoxHeight;
-    if (!missingDimAttr) {
-        parsedDim.width = parseUnit(widthAttr);
-        parsedDim.height = parseUnit(heightAttr);
-        scaleX = parsedDim.width / viewBoxWidth;
-        scaleY = parsedDim.height / viewBoxHeight;
-    }
-    else {
-        parsedDim.width = viewBoxWidth;
-        parsedDim.height = viewBoxHeight;
-    }
-
-    // default is to preserve aspect ratio
-    preserveAspectRatio = parsePreserveAspectRatioAttribute(preserveAspectRatio);
-    if (preserveAspectRatio.alignX !== 'none') {
-        //translate all container for the effect of Mid, Min, Max
-        if (preserveAspectRatio.meetOrSlice === 'meet') {
-            scaleY = scaleX = (scaleX > scaleY ? scaleY : scaleX);
-            // calculate additional translation to move the viewbox
-        }
-        if (preserveAspectRatio.meetOrSlice === 'slice') {
-            scaleY = scaleX = (scaleX > scaleY ? scaleX : scaleY);
-            // calculate additional translation to move the viewbox
-        }
-        widthDiff = parsedDim.width - viewBoxWidth * scaleX;
-        heightDiff = parsedDim.height - viewBoxHeight * scaleX;
-        if (preserveAspectRatio.alignX === 'Mid') {
-            widthDiff /= 2;
-        }
-        if (preserveAspectRatio.alignY === 'Mid') {
-            heightDiff /= 2;
-        }
-        if (preserveAspectRatio.alignX === 'Min') {
-            widthDiff = 0;
-        }
-        if (preserveAspectRatio.alignY === 'Min') {
-            heightDiff = 0;
-        }
-    }
-
-    if (scaleX === 1 && scaleY === 1 && minX === 0 && minY === 0 && x === 0 && y === 0) {
-        return parsedDim;
-    }
-    if ((x || y) && element.parentNode.nodeName !== '#document') {
-        translateMatrix = ' translate(' + parseUnit(x) + ' ' + parseUnit(y) + ') ';
-    }
-
-    matrix = translateMatrix + ' matrix(' + scaleX +
-        ' 0' +
-        ' 0 ' +
-        scaleY + ' ' +
-        (minX * scaleX + widthDiff) + ' ' +
-        (minY * scaleY + heightDiff) + ') ';
-    // seems unused.
-    // parsedDim.viewboxTransform = fabric.parseTransformAttribute(matrix);
-    if (element.nodeName === 'svg') {
-        el = element.ownerDocument.createElementNS(svgNS, 'g');
-        // element.firstChild != null
-        while (element.firstChild) {
-            el.appendChild(element.firstChild);
-        }
-        element.appendChild(el);
-    }
-    else {
-        el = element;
-        el.removeAttribute('x');
-        el.removeAttribute('y');
-        matrix = el.getAttribute('transform') + matrix;
-    }
-    el.setAttribute('transform', matrix);
-    return parsedDim;
-}
-
-export function hasAncestorWithNodeName(element, nodeName) {
-    while (element && (element = element.parentNode)) {
-        if (element.nodeName && nodeName.test(element.nodeName.replace('svg:', ''))
-            && !element.getAttribute('instantiated_by_use')) {
-            return true;
-        }
-    }
-    return false;
-}
