+++ src/parser/elementMatchesRule.ts
@@ -1,0 +1,18 @@
+//@ts-nocheck
+
+import { selectorMatches } from "./selectorMatches";
+import { doesSomeParentMatch } from "./doesSomeParentMatch";
+
+/**
+ * @private
+ */
+
+export function elementMatchesRule(element, selectors) {
+    var firstMatching, parentMatching = true;
+    //start from rightmost selector.
+    firstMatching = selectorMatches(element, selectors.pop());
+    if (firstMatching && selectors.length) {
+        parentMatching = doesSomeParentMatch(element, selectors);
+    }
+    return firstMatching && parentMatching && (selectors.length === 0);
+}
