+++ src/parser/getMultipleNodes.ts
@@ -1,0 +1,11 @@
+//@ts-nocheck
+
+export function getMultipleNodes(doc, nodeNames) {
+    var nodeName, nodeArray = [], nodeList, i, len;
+    for (i = 0, len = nodeNames.length; i < len; i++) {
+        nodeName = nodeNames[i];
+        nodeList = doc.getElementsByTagName(nodeName);
+        nodeArray = nodeArray.concat(Array.prototype.slice.call(nodeList));
+    }
+    return nodeArray;
+}
