+++ src/parser/parseStyleObject.ts
@@ -1,0 +1,15 @@
+//@ts-nocheck
+
+export function parseStyleObject(style, oStyle) {
+    var attr, value;
+    for (var prop in style) {
+        if (typeof style[prop] === 'undefined') {
+            continue;
+        }
+
+        attr = prop.toLowerCase();
+        value = style[prop];
+
+        oStyle[attr] = value;
+    }
+}
