+++ src/parser/translateMatrix.ts
@@ -1,0 +1,8 @@
+//@ts-nocheck
+
+export function translateMatrix(matrix, args) {
+    matrix[4] = args[0];
+    if (args.length === 2) {
+        matrix[5] = args[1];
+    }
+}
